Move `ItemListing` to `proc_util` crate
